Reference

PrincipalMomentAnalysis.pma — Function

pma(A, G; nsv=6)

Computes the Principal Moment Analysis of the matrix A (variables × samples) using the sample adjacency graph G. Set nsv to control the number of singular values and vectors returned. Returns a PMA struct.

See also PMA.

PrincipalMomentAnalysis.PMA — Type

PMA <: Factorization

The output of pma representing the Principal Moment Analysis factorization of a matrix A. If F::PMA is the factorization object, U, S, V and Vt can be obtained via F.U, F.S, F.V and F.Vt such as A ≈ U * Diagonal(S) * Vt'.

See also pma.

PrincipalMomentAnalysis.normalizemean! — Function

normalizemean!(X)

In place removal of the variable mean from the P×N matrix X where P is the number of variables and N is the number of samples.

PrincipalMomentAnalysis.normalizemean — Function

normalizemean(X)

Remove the variable mean from the P×N matrix X where P is the number of variables and N is the number of samples.

PrincipalMomentAnalysis.normalizemeanstd! — Function

normalizemeanstd!(X)

In place normalization of variables to be mean zero and standard deviation one of the P×N matrix X where P is the number of variables and N is the number of samples.

PrincipalMomentAnalysis.normalizemeanstd — Function

normalizemeanstd!(X)

Normalize variables to be mean zero and standard deviation one in the P×N matrix X where P is the number of variables and N is the number of samples.