Reference

PrincipalMomentAnalysis.pma — Function

pma(A, G; nsv=6)

Computes the Principal Moment Analysis of the matrix A (variables × samples) using the sample simplex graph G. Each column in G is a boolean vector representing one simplex. True means that a vertex is part of the simplex and false that it is not. Set nsv to control the number of singular values and vectors returned. Returns a PMA struct.

See also PMA.

PrincipalMomentAnalysis.PMA — Type

PMA <: Factorization

The output of pma representing the Principal Moment Analysis factorization of a matrix A. If F::PMA is the factorization object, U, S, V and Vt can be obtained via F.U, F.S, F.V and F.Vt such as A ≈ U * Diagonal(S) * Vt'.

See also pma.

PrincipalMomentAnalysis.normalizemean! — Function

normalizemean!(A)

In place removal of the variable mean from the P×N matrix A where P is the number of variables and N is the number of samples.

PrincipalMomentAnalysis.normalizemean — Function

normalizemean(A)

Remove the variable mean from the P×N matrix A where P is the number of variables and N is the number of samples.

PrincipalMomentAnalysis.normalizemeanstd! — Function

normalizemeanstd!(A)

In place normalization of variables to be mean zero and standard deviation one of the P×N matrix A where P is the number of variables and N is the number of samples.

PrincipalMomentAnalysis.normalizemeanstd — Function

normalizemeanstd!(A)

Normalize variables to be mean zero and standard deviation one in the P×N matrix A where P is the number of variables and N is the number of samples.

PrincipalMomentAnalysis.groupsimplices — Function

groupsimplices(groupby::AbstractVector)

Create simplex graph connecting elements with identical values in the groupby vector.

Examples

julia> G = groupsimplices(["A","A","B","C","B"])
5×5 BitArray{2}:
 1  1  0  0  0
 1  1  0  0  0
 0  0  1  0  1
 0  0  0  1  0
 0  0  1  0  1

PrincipalMomentAnalysis.timeseriessimplices — Function

timeseriessimplices(time::AbstractVector; groupby::AbstractVector)

Create simplex graph connecting elements adjacent in time. In case of ties, All elements at a unique timepoint will be connected to the all elements at the previous, current and next timepoints. If groupby is specified, the elements are first divided by group, and then connected by time.

Examples

julia> G = timeseriessimplices([0.5, 1.0, 4.0])
3×3 BitArray{2}:
 1  1  0
 1  1  1
 0  1  1

julia> G = timeseriessimplices([0.5, 1.0, 1.0, 4.0])
4×4 BitArray{2}:
 1  1  1  0
 1  1  1  1
 1  1  1  1
 0  1  1  1

julia> G = timeseriessimplices([2, 4, 6, 2, 4, 8]; groupby=["A","A","A","B","B","B"])
6×6 BitArray{2}:
 1  1  0  0  0  0
 1  1  1  0  0  0
 0  1  1  0  0  0
 0  0  0  1  1  0
 0  0  0  1  1  1
 0  0  0  0  1  1

PrincipalMomentAnalysis.neighborsimplices — Function

neighborsimplices(A::AbstractMatrix; k, r, dim, symmetric, normalizedist, groupby)

Create simplex graph connecting nearest neighbor samples.

Inputs

A: Data matrix (variables × samples).

k: Number of nearest neighbors to connect. Default: 0.

r: Connected all neighbors with disctance ≤r. Default: 0.0.

dim: Reduce the dimension to dim before computing distances. Useful to reduce noise. Default: Disabled.

symmetric: Make the simplex graph symmetric. Default: false.

normalizedist: Normalize distances to the scale [0.0,1.0]. Affects the r parameter. Default: true.

groupby: Only connected samples within the specified groups. Default: Disabled.

Examples

julia> G = neighborsimplices([0 0 2 2; 0 1 1 0]; k=1)
4×4 BitArray{2}:
 1  1  0  0
 1  1  0  0
 0  0  1  1
 0  0  1  1

julia> G = neighborsimplices([0 0 2 2; 0 1 1 0]; r=0.9)
4×4 BitArray{2}:
 1  1  0  1
 1  1  1  0
 0  1  1  1
 1  0  1  1

PrincipalMomentAnalysis.neighborsimplices2 — Function

neighborsimplices2(D2; k, r, symmetric, normalizedist, groupby)

Create simplex graph connecting nearest neighbors in given symmetric matrix where element i,j equals the squared distance between samples i and j.

Inputs

D2: Matrix of squared distances.

k: Number of nearest neighbors to connect. Default: 0.

r: Connected all neighbors with disctance ≤r. Default: 0.0.

symmetric: Make the simplex graph symmetric. Default: false.

normalizedist: Normalize distances to the scale [0.0,1.0]. Affects the r parameter. Default: true.

groupby: Only connected samples within the specified groups. Default: Disabled.